changed return format and added environmental variables
